# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Kazoot
  description: |-
    An application
  termsOfService: https://kazoot.no/terms-of-service
  contact:
    email: vkbugge@hotmail.com
  license:
    name: MIT License
    url: "https://opensource.org/licenses/MIT"
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: https://api.kazoot.no
...
paths:
...
  /user:
    post:
      tags:
        - User
      description: Register a user (create a user)
      operationId: registerUser
      parameters:
        - name: anonymous
          description: If the user should be anonymous
          in: query
          required: false
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: User already exists
    put:
      tags:
        - User
      description: Update the information of a user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: User already exists
  /user/{id}:
    get:
      tags:
        - User
      description: Get a specific user
      operationId: fetchUser
      parameters:
        - name: id
          description: The ID of the user
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags:
        - User
      description: Delete a specific User
      operationId: deleteUser
      parameters:
        - name: id
          description: The ID of the user
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successful response
  /user/session:
    post:
      tags:
        - User Session
      description: Log in with credentials
      operationId: login
      requestBody:
        description: Login credentials
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          headers:
            Set-Cookie:
              description: Cookie to be used for authentication.
              schema:
                type: string
                example: "Authorization=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1NzYyMjQxM"
        "401":
          description: Bad credentials
    put:
      tags:
        - User Session
      description: Refresh token
      operationId: refreshToken
      responses:
        "201":
          description: Successful operation
          headers:
            Set-Cookie:
              description: Cookie to be used for authentication.
              schema:
                type: string
                example: "Authorization=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InRlc3QiLCJpYXQiOjE1NzYyMjQxM"
        "401":
          description: Bad credentials
    get:
      tags:
        - User Session
      description: Get logged in user
      operationId: loggedInUser
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Bad credentials
    delete:
      tags:
        - User Session
      description: Log out User
      operationId: logout
      responses:
        "204":
          description: Successful operation
  /quiz:
    post:
      tags:
        - Quiz
      description: Create a new quiz
      operationId: createQuiz
      requestBody:
        description: Quiz creation information
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCreate"
            example:
              {
                "title": "Biology: Chapter 1",
                "description": "Quiz about the human body.",
                "difficulty": 8,
                "categories": [],
              }
      responses:
        "201":
          description: Successfully created Quiz
        "401":
          description: Not logged in
    get:
      tags:
        - Quiz
      description: Get first page of quizzes
      operationId: getQuizzes
      parameters:
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: textSearch
          schema:
            type: string
            maxLength: 128
        - in: query
          name: minDifficulty
          schema:
            type: integer
        - in: query
          name: maxDifficulty
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: array
            items:
              type: string
              format: uuid
        - in: query
          name: creator
          schema:
            type: string
            format: uuid
        - in: query
          name: collaborator
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully got the first page of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizOverview"
  /quiz/{id}:
    get:
      tags:
        - Quiz
      description: Get all information about a quiz
      operationId: getQuiz
      parameters:
        - name: id
          description: The ID of a quiz
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "Successfully fetched the quiz"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    put:
      tags:
        - Quiz
      description: Update quiz info
      operationId: updateQuiz
      parameters:
        - name: id
          description: The ID of a quiz
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: The updated quiz info
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCreate"
      responses:
        "200":
          description: "Successful update"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
...
components:
  schemas:
    Image:
      type: object
      required:
        - id
        - path
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    UserUpdate:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    UserCreate:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    Category:
      type: object
      required:
        - id
        - name
        - color
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 32
        color:
          type: string
          minLength: 4
          maxLength: 7
...
    Feedback:
      type: object
      required:
        - id
        - quiz_id
        - username
        - feedback
        - created_at
      properties:
        id:
          type: integer
        quiz_id:
          type: string
          format: uuid
        username:
          type: string
        feedback:
          type: string
          maxLength: 1024
        created_at:
          type: string
          format: date-time
...
    Question:
      type: object
      required:
        - id
        - question
        - options
        - image
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
          maxLength: 128
        answer:
          type: string
        image:
          $ref: "#/components/schemas/Image"
        options:
          type: array
          maxLength: 4
          items:
            type: string
    QuizCreate:
      type: object
      required:
        - title
        - description
        - difficulty
        - categories
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        difficulty:
          type: integer
        categories:
          type: array
          items:
            type: string
            format: uuid
    QuizOverview:
      type: object
      required:
        - id
        - title
        - description
        - categories
        - creator
        - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 5
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        creator:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
          format: date-time
    Quiz:
      type: object
      required:
        - id
        - title
        - description
        - categories
        - feedbacks
        - questions
        - creator
        - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 5
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        feedbacks:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        creator:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
          format: date-time
...